Add include all option for variable
It is now possible to add an include all values option for a variable.
If this option is activated and selected, the query which uses the
variable is run with all values. For Prometheus we are passing ".*" as
value to the query which is the same as using "value1|value2|value3".