Fix dashboard refresh
The dashboard wasn't refreshed in the given refresh time, because the
start and end time of the interval were not updated befor the refresh
was triggered.

Also the refresh function isn't executed if the modal is closed without
action (ESC), because the refresh function closed the modal every time.
This was a bit frustrating specially for small time intervals.

Instead the dashboard is refreshed when the modal is closed.