Add explore mode
Add a new "explore" command option, to enter the explore mode of dash.
The explore mode can be used to directly query your datasource, with out
the need to build a dashboard.

The explore mode supports a "query" flag, which let you pass a query to
the explore mode.

To add the support mode a new method for the datasources was added. Also
the linechart component was adjusted to support better rendering of the
legend in the new mode. Instead of the Variables the query is shown in
the statusbar.